+++ app/models.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+class PromptRequest(BaseModel):
+    prompt: str
+    max_tokens: int = 256
+    temperature: float = 0.7
+    top_p: float = 0.95
